VIITE-2011:
	- When roadway_number changes, we should not add a calibration point
